+++ src/app/(main)/[productType]/components/SelectedProductImages.tsx
@@ -1,4 +1,4 @@
-import ReviewImagesSheet from '@/components/PDP/components/ReviewImagesSheet';
+import ReviewImages from '@/components/PDP/components/ReviewImages';
 import { Dialog, DialogContent, DialogTrigger } from '@/components/ui/dialog';
 import Image from 'next/image';
 import { FaCamera } from 'react-icons/fa';
@@ -88,7 +88,7 @@ const MoreImages = () => (
         </div>
       </DialogTrigger>
       <DialogContent className="flex min-h-[65vh] flex-col items-center lg:min-w-[77vw] lg:max-w-[80%]">
-        <ReviewImagesSheet />
+        <ReviewImages />
       </DialogContent>
     </Dialog>
   </div>
