+++ src/components/PDP/components/EditVehiclePopover.tsx
@@ -23,7 +23,7 @@ export default function EditVehiclePopover({
         <Drawer.Trigger className="flex h-full items-center justify-between text-left text-base text-[#1A1A1A]">
           <div className="my-2 border-l-2 border-l-[#C8C7C7] pl-8 pr-8">
             <p>Your Vehicle</p>
-            <h2 className=" whitespace-nowrap text-[18px] font-[500] capitalize leading-[22px] text-[#1A1A1A]">
+            <h2 className=" text-[18px] font-[500] capitalize leading-[22px] text-[#1A1A1A]">
               {fullProductName}
             </h2>
           </div>
