+++ src/components/PDP/components/ReviewCard.tsx
@@ -67,7 +67,7 @@ export default function ReviewCard({
         {/* {!tabsCard && (
           <HelpfulSection numberOfHelpful={review?.helpful as string} />
         )} */}
-        {/* <ReviewCardImages reviewImages={review?.review_image} /> */}
+        <ReviewCardImages reviewImages={review?.review_image} />
       </div>
     </div>
   );
