+++ src/components/PDP/components/ReviewSheet.tsx
@@ -37,7 +37,7 @@ export default function ReviewSheet({ seeMore }: { seeMore?: boolean }) {
         <SheetHeader draggable={false}>
           <SheetTitle className="sticky top-0 z-[100] mx-4 flex justify-between bg-white">
             <SheetClose className="fixed right-0 z-[400] mr-[16px] flex items-center py-[4px]">
-              <button
+              <div
                 id="CloseModalButton"
                 className=" mt-[17px] justify-center rounded-full bg-gray-200 p-[5px] "
                 onClick={() => {
@@ -45,7 +45,7 @@ export default function ReviewSheet({ seeMore }: { seeMore?: boolean }) {
                 }}
               >
                 <IoClose className="h-[24px] w-[24px]" />
-              </button>
+              </div>
             </SheetClose>
           </SheetTitle>
         </SheetHeader>
