+++ src/components/PDP/components/SuggestedProducts.tsx
@@ -58,8 +58,6 @@ export default function SuggestedProducts() {
     },
   ];
 
-  console.log(products);
-
   return (
     <section id="you-may-also-like-section" className="px-4 pb-[20px]">
       <h1 className="font-[700] uppercase lg:text-[32px] lg:font-[900]">
